pyproject.toml
[project]
name = "alien"
version = "0.1.0"
description = "O-1 Visa Readiness Analyzer - AI-powered resume analysis for extraordinary ability visa assessment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "jinja2",
    "python-multipart",
    "pydantic",
    "orjson",
    "openai",
    "pymupdf",
    "python-docx",
    "python-dotenv",
    "websockets",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]
